refactor: improve the inconvenience of handling font in the community through two types of spec
