+++ packages/blocks/package.json
@@ -26,7 +26,6 @@
     "@types/hast": "^3.0.4",
     "@types/mdast": "^4.0.3",
     "@types/sortablejs": "^1.15.8",
-    "@types/webfontloader": "^1.6.38",
     "buffer": "^6.0.3",
     "date-fns": "^3.6.0",
     "dompurify": "^3.0.11",
@@ -56,7 +55,6 @@
     "shiki": "^1.2.0",
     "sortablejs": "^1.15.2",
     "unified": "^11.0.4",
-    "webfontloader": "^1.6.28",
     "zod": "^3.22.4"
   },
   "exports": {
