+++ packages/blocks/src/embed-synced-doc-block/embed-synced-doc-block.ts
@@ -17,12 +17,12 @@ import { EmbedBlockElement } from '../_common/embed-block-helper/embed-block-ele
 import { REFERENCE_NODE } from '../_common/inline/presets/nodes/consts.js';
 import { matchFlavours } from '../_common/utils/model.js';
 import { getThemeMode } from '../_common/utils/query.js';
+import type { NoteBlockModel } from '../note-block/note-model.js';
+import type { RootBlockComponent } from '../root-block/index.js';
 import {
   EdgelessEditorBlockSpecs,
   PageEditorBlockSpecs,
-} from '../_specs/_specs.js';
-import type { NoteBlockModel } from '../note-block/note-model.js';
-import type { RootBlockComponent } from '../root-block/index.js';
+} from '../specs/index.js';
 import { Bound } from '../surface-block/utils/bound.js';
 import type { EmbedSyncedDocCard } from './components/embed-synced-doc-card.js';
 import type { EmbedSyncedDocModel } from './embed-synced-doc-model.js';
