+++ packages/blocks/src/index.ts
@@ -72,7 +72,6 @@ export {
 } from './_common/utils/index.js';
 export { createDefaultDoc } from './_common/utils/init.js';
 export { getThemeMode, isInsidePageEditor } from './_common/utils/query.js';
-export * from './_specs/_specs.js';
 export * from './attachment-block/index.js';
 export * from './bookmark-block/index.js';
 export * from './code-block/index.js';
@@ -102,11 +101,12 @@ export { CopilotSelectionController } from './root-block/edgeless/controllers/to
 export type { EdgelessModel } from './root-block/edgeless/type.js';
 export * from './root-block/index.js';
 export * from './schemas.js';
+export * from './specs/index.js';
+export * from './surface-block/consts.js';
 export {
   Bound,
   BrushElementModel,
   CanvasElementType,
-  CanvasTextFonts,
   ConnectorElementModel,
   ConnectorMode,
   ElementModel,
