+++ packages/blocks/src/specs/base/index.ts
@@ -1,0 +1,3 @@
+export * from './edgeless-root.js';
+export * from './general.js';
+export * from './page-root.js';
