+++ packages/blocks/src/specs/index.ts
@@ -1,0 +1,3 @@
+export * from './base/index.js';
+export * from './specs.js';
+export * from './utils/index.js';
