+++ packages/presets/src/ai/messages/text.ts
@@ -1,11 +1,13 @@
 import { type EditorHost } from '@blocksuite/block-std';
-import { type AffineAIPanelWidgetConfig } from '@blocksuite/blocks';
+import {
+  type AffineAIPanelWidgetConfig,
+  PageEditorBlockSpecs,
+} from '@blocksuite/blocks';
 import type { Doc } from '@blocksuite/store';
 import { css, html, LitElement, nothing, type PropertyValues } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 import { keyed } from 'lit/directives/keyed.js';
 
-import { CustomPageEditorBlockSpecs } from '../utils/custom-specs.js';
 import { markDownToDoc } from '../utils/markdown-utils.js';
 
 @customElement('ai-answer-text')
@@ -79,10 +81,7 @@ export class AIAnswerText extends LitElement {
     return html`
       <div class="ai-answer-text-container">
         <div class="ai-answer-text-editor affine-page-viewport">
-          ${this.host.renderSpecPortal(
-            this._previewDoc,
-            CustomPageEditorBlockSpecs
-          )}
+          ${this.host.renderSpecPortal(this._previewDoc, PageEditorBlockSpecs)}
         </div>
       </div>
     `;
