+++ packages/presets/src/index.ts
@@ -2,6 +2,7 @@ export * from './ai/index.js';
 export * from './editors/index.js';
 export * from './fragments/index.js';
 export * from './helpers/index.js';
+export * from './specs/index.js';
 
 const env =
   typeof globalThis !== 'undefined'
