+++ packages/presets/src/specs/index.ts
@@ -1,0 +1,2 @@
+export * from './affine.js';
+export * from './community.js';
